Remove defaults from Timer object, add them to beginTimer function
I considered removing default values entirely, as with TimerView in
the previous commit, but then I'd have to have fill in all values
including in Index.js and TimerView.js, which would be redundant.
Another option would be to remove the object entirely and just
pass ALL the values as parameters between all the functions,which
would be even more verbose and redundant.

Since beginTimer is only called with non-default values inside
Timer.js, and only called with default ones from outside, I think
this is a good solution.