Update UI and UX
Since I've been figuring out how to make things work as I go,
the UI and UX have been a kludgey mess from the beginning.
Originally with this commit I had simply intended to  to update the
settings menu to fit the screen, but as I got into it many related
issues became apparent.

With this update the UI is simpler, the UX is more
intuitive (I think), and the code is *much* cleaner. I expect it
will be easier to work with from now on as well.