Remove storage functionality and code
I've hit a wall adding this functionality, and it's killing my
motivation. I've decided it's time to move on. I've kept the code
in the storage branch, so I can go back to it easily if the mood
strikes. Or, ya know, I could just grab it from an earlier commit.